GPU: Clean up utilization of multi-lane TPC clusterizer, some tuning to GPU memory

+++ GPU/GPUTracking/gpucf/GPUTPCClusterFinderKernels.cxx
@@ -93,7 +93,7 @@ GPUd() void GPUTPCClusterFinderKernels::Thread<GPUTPCClusterFinderKernels::compa
 {
   gpucf::compactDigit(get_num_groups(0), get_local_size(0), get_group_id(0), get_local_id(0), smem, in, out, clusterer.mPisPeak, clusterer.mPbuf + (iBuf - 1) * clusterer.mBufSize, clusterer.mPbuf + iBuf * clusterer.mBufSize);
   unsigned int lastId = get_global_size(0) - 1;
-  if (get_global_id(0) == lastId) {
+  if ((unsigned int)get_global_id(0) == lastId) {
     if (stage) {
       clusterer.mPmemory->nClusters = clusterer.mPbuf[lastId];
     } else {
